epistula/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic[email]>=2.0.0

# Database
sqlalchemy>=2.0.35
psycopg[binary]>=3.2.0
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests>=2.31.0
minio>=7.2.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.5
pytest-xdist>=3.5.0
httpx==0.27.2
asgi-lifespan>=2.1.0
